+++ tap-snapshots/lib-test-TAP.test.js
@@ -5,66 +5,16 @@
  * Make sure to inspect the output below.  Do not ignore changes!
  */
 'use strict'
-exports[`lib/test TAP \`errorHandler\` middleware handles a standard error. > Koa app error event. 1`] = `
-Error {
-  "expose": false,
-  "statusCode": 500,
-  "status": 500
-}
-`
-
 exports[`lib/test TAP \`errorHandler\` middleware handles a standard error. > Response body. 1`] = `
 {"errors":[{"message":"Internal Server Error"}]}
-`
-
-exports[`lib/test TAP \`errorHandler\` middleware handles a HTTP error. > Koa app error event. 1`] = `
-ForbiddenError {
-  "message": "Test."
-}
 `
 
 exports[`lib/test TAP \`errorHandler\` middleware handles a HTTP error. > Response body. 1`] = `
 {"errors":[{"message":"Test."}]}
 `
 
-exports[`lib/test TAP \`execute\` middleware options missing. > Creation error. 1`] = `
-InternalServerError {
-  "message": "GraphQL execute middleware options missing."
-}
-`
-
-exports[`lib/test TAP \`execute\` middleware options not an object. > Creation error. 1`] = `
-InternalServerError {
-  "message": "GraphQL execute middleware options must be an object."
-}
-`
-
-exports[`lib/test TAP \`execute\` middleware options invalid. > Creation error. 1`] = `
-InternalServerError {
-  "message": "GraphQL execute middleware options invalid: \`invalid1\`, \`invalid2\`."
-}
-`
-
-exports[`lib/test TAP \`execute\` middleware option \`override\` options invalid. > Koa app error event. 1`] = `
-InternalServerError {
-  "message": "GraphQL execute middleware \`override\` option return options invalid: \`invalid\`, \`override\`."
-}
-`
-
 exports[`lib/test TAP \`execute\` middleware option \`override\` options invalid. > Response body. 1`] = `
 {"errors":[{"message":"Internal Server Error"}]}
-`
-
-exports[`lib/test TAP \`execute\` middleware option \`override\` not a function. > Creation error. 1`] = `
-InternalServerError {
-  "message": "GraphQL execute middleware \`override\` option must be a function."
-}
-`
-
-exports[`lib/test TAP \`execute\` middleware option \`override\` not an object. > Koa app error event. 1`] = `
-InternalServerError {
-  "message": "GraphQL execute middleware options must be an object, or an object promise."
-}
 `
 
 exports[`lib/test TAP \`execute\` middleware option \`override\` not an object. > Response body. 1`] = `
@@ -99,148 +49,36 @@ exports[`lib/test TAP \`execute\` middleware option \`fieldResolver\` override u
 {"data":{"test":"fieldResolverOverridden"}}
 `
 
-exports[`lib/test TAP \`execute\` middleware option \`schema\` not a GraphQLSchema instance. > Creation error. 1`] = `
-InternalServerError {
-  "message": "GraphQL schema is required and must be a \`GraphQLSchema\` instance."
-}
-`
-
-exports[`lib/test TAP \`execute\` middleware option \`schema\` not a GraphQLSchema instance override. > Koa app error event. 1`] = `
-InternalServerError {
-  "message": "GraphQL schema is required and must be a \`GraphQLSchema\` instance."
-}
-`
-
 exports[`lib/test TAP \`execute\` middleware option \`schema\` not a GraphQLSchema instance override. > Response body. 1`] = `
 {"errors":[{"message":"Internal Server Error"}]}
-`
-
-exports[`lib/test TAP \`execute\` middleware option \`schema\` invalid GraphQL. > Creation error. 1`] = `
-InternalServerError {
-  "message": "GraphQL schema validation errors.",
-  "graphqlErrors": [
-    {
-      "message": "Query root type must be provided."
-    }
-  ]
-}
-`
-
-exports[`lib/test TAP \`execute\` middleware option \`schema\` invalid GraphQL override. > Koa app error event. 1`] = `
-InternalServerError {
-  "message": "GraphQL schema validation errors.",
-  "graphqlErrors": [
-    {
-      "message": "Query root type must be provided."
-    }
-  ]
-}
 `
 
 exports[`lib/test TAP \`execute\` middleware option \`schema\` invalid GraphQL override. > Response body. 1`] = `
 {"errors":[{"message":"Internal Server Error"}]}
 `
 
-exports[`lib/test TAP Request body missing due to absent body parser middleware. > Koa app error event. 1`] = `
-InternalServerError {
-  "message": "Request body missing."
-}
-`
-
 exports[`lib/test TAP Request body missing due to absent body parser middleware. > Response body. 1`] = `
 {"errors":[{"message":"Internal Server Error"}]}
-`
-
-exports[`lib/test TAP Request body invalid. > Koa app error event. 1`] = `
-BadRequestError {
-  "message": "Request body must be a JSON object."
-}
 `
 
 exports[`lib/test TAP Request body invalid. > Response body. 1`] = `
 {"errors":[{"message":"Request body must be a JSON object."}]}
 `
 
-exports[`lib/test TAP Operation field \`query\` missing. > Koa app error event. 1`] = `
-BadRequestError {
-  "message": "GraphQL operation field \`query\` missing."
-}
-`
-
 exports[`lib/test TAP Operation field \`query\` missing. > Response body. 1`] = `
 {"errors":[{"message":"GraphQL operation field \`query\` missing."}]}
-`
-
-exports[`lib/test TAP Operation field \`query\` invalid. > Koa app error event. 1`] = `
-BadRequestError {
-  "message": "GraphQL query syntax error: Syntax Error: Unexpected ["
-}
 `
 
 exports[`lib/test TAP Operation field \`query\` invalid. > Response body. 1`] = `
 {"errors":[{"message":"GraphQL query syntax error: Syntax Error: Unexpected ["}]}
 `
 
-exports[`lib/test TAP Operation field \`variables\` invalid. > Koa app error event. 1`] = `
-BadRequestError {
-  "message": "GraphQL operation field invalid: Variables must be provided as an Object where each property is a variable value. Perhaps look to see if an unparsed JSON string was provided."
-}
-`
-
 exports[`lib/test TAP Operation field \`variables\` invalid. > Response body. 1`] = `
 {"errors":[{"message":"GraphQL operation field invalid: Variables must be provided as an Object where each property is a variable value. Perhaps look to see if an unparsed JSON string was provided."}]}
 `
 
-exports[`lib/test TAP Operation field \`query\` validation errors. > Koa app error event. 1`] = `
-BadRequestError {
-  "message": "GraphQL query validation errors.",
-  "graphqlErrors": [
-    {
-      "message": "Cannot query field \\"wrongOne\\" on type \\"Query\\".",
-      "locations": [
-        {
-          "line": 1,
-          "column": 9
-        }
-      ]
-    },
-    {
-      "message": "Cannot query field \\"wrongTwo\\" on type \\"Query\\".",
-      "locations": [
-        {
-          "line": 1,
-          "column": 19
-        }
-      ]
-    }
-  ]
-}
-`
-
 exports[`lib/test TAP Operation field \`query\` validation errors. > Response body. 1`] = `
 {"errors":[{"message":"Cannot query field \\"wrongOne\\" on type \\"Query\\".","locations":[{"line":1,"column":9}]},{"message":"Cannot query field \\"wrongTwo\\" on type \\"Query\\".","locations":[{"line":1,"column":19}]}]}
-`
-
-exports[`lib/test TAP GraphQL resolver error. > Koa app error event. 1`] = `
-Error {
-  "expose": true,
-  "statusCode": 200,
-  "status": 200,
-  "graphqlErrors": [
-    {
-      "message": "Resolver error.",
-      "locations": [
-        {
-          "line": 1,
-          "column": 3
-        }
-      ],
-      "path": [
-        "test"
-      ]
-    }
-  ]
-}
 `
 
 exports[`lib/test TAP GraphQL resolver error. > Response body. 1`] = `